COMMERCE-4667 commerce-product-api moved classes from commerce-product-definition-web
